callback wrapped in suspense
